pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "canarygan"
version = "0.0.1"
dependencies = [
  "torch>=2.0.1,<3.0.0",
  "torchaudio>=2.0.2,<3.0.0",
  "torchvision>=0.15.2,<1.0.0",
  "torchmetrics>=1.1.1,<2.0.0",
  "lightning>=2.0.8,<3.0.0",
  "matplotlib==3.7.2",
  "pandas==2.1.0",
  "librosa==0.10.1",
  "pydub==0.25.1",
  "numpy<=2.0.0",
  "scipy<=2.0.0",
  "tensorboardX==2.6.2.2",
  "tensorboard==2.14.0",
  "seaborn==0.13.0",
  "scikit-learn==1.3.0",
  "pyyaml>=6.0.1,<=7.0.0",
  "rich>=13.5.2",
  "reservoirpy>=0.3.10,<0.4.0",
  "joblib==1.3.2",
  "dill==0.3.7",
  "umap-learn==0.5.5",
]
requires-python = ">=3.9,<=3.12"
authors = [
  {name = "Nathan Trouvain", email = "ntrouvain@ensc.fr"},
]
description = "A Pytorch reimplementation of Pagliarini et al. (2021) 'What does the Canary Say? Low-Dimensional GAN Applied to Birdsong'"
readme = "README.md"
license = {file = "LICENSE"}

[project.urls]
Homepage = "https://gihtub.com/ntrouvain/canarygan.git"
Repository = "https://github.com/ntrouvain/canarygan.git"

[project.scripts]
spam-cli = "spam:main_cli"

